Add more code to facilitate actor implementation
Added obtain() to Context
Added startGuardAsync(), in anticipation of adding guards.
Created HttpOperator superclass.
Moved logRestXxx() from Util to HttpOperator.
Updated SDNC with new HttpOperator.
Added actor.test to facilitate testing of actors.
Changed timeoutSec from long to int in various places.
Made a couple of methods public to support junit testing.
Split Operator and Operation.
Simplified HttpOperation, eliminating dual nature.

Issue-ID: POLICY-1625

+++ models-interactions/model-actors/actorServiceProvider/src/main/java/org/onap/policy/controlloop/actorserviceprovider/parameters/HttpParams.java
@@ -53,7 +53,7 @@ public class HttpParams {
      */
     @Min(0)
     @Builder.Default
-    private long timeoutSec = 0;
+    private int timeoutSec = 0;
 
 
     /**
